Fixes
- Enemy Melee controller stats werent being reset properly since changes, causing amplified stats after 2nd lifetime from Pooling.
- Explosions were not going back into their pool when they had no explosion radius
- Made the Tutorial planet give more credits per wave to help speed it up
- Changed BatchMult to not be a float instead of an int, can't beleive that wasn't done already!  Lowered tutorial batchMultStart from 2 to 1.5
- Fixed CritHitFX not repooling

+++ Mutant Mayhem/Assets/_Scripts/Planets/PlanetSO.cs
@@ -29,7 +29,7 @@ public class PlanetSO : ScriptableObject
     public float wavesTillAddIndex = 1;
     public int creditsPerWave = 150; // Additive bonus (waveIndex*creditsPerWave)
     public float timeBetweenWavesBase = 90; // Base amount of day-time
-    public int batchMultiplierStart = 5; // Starting batch multiplier for each Subwave
+    public float batchMultiplierStart = 5; // Starting batch multiplier for each Subwave
     public float subwaveDelayMultStart = 1f; // Time between subwaves
     public int spawnRadiusBuffer = 16;
     public int batchMultGrowthTime = 2;
